pyproject.toml
[tool.poetry]
name = "repl_nix_VAI"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "1.10.7"
langchain = "0.0.170"
python-dotenv = "1.0.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
